experiments/fraud_mini/requirements.txt
tensorflow-gpu==1.14
configargparse
tqdm
matplotlib
sklearn
